Can now add MovieDirectors (Many-To-Many join table addition)

+++ CoolestMovieAPI/CoolestMovieAPI/DTO/MovieDTO.cs
@@ -17,7 +17,7 @@ namespace CoolestMovieAPI.DTO
         [Required]
         public string MovieTitle { get; set; }
         public TimeSpan MovieLength { get; set; }
-        public int MovieRating { get; set; }
+        public double MovieRating { get; set; }
         public string MovieDescription { get; set; }
         public int MovieReleaseYear { get; set; }
         public ICollection<MovieDirectorDTO> MovieDirectors { get; set; }
